requirements.txt
openai>=1.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
ebooklib>=0.18
beautifulsoup4>=4.11.0